requirements.txt
torch==1.11.0+cu113
numpy==1.22.4
PyYAML==6.0.2
scipy==1.10.1
tqdm==4.61.2